+++ kernel/sysfs.h
@@ -1,0 +1,9 @@
+/*
+ * sysfs.h - Kernel object information filesystem
+ */
+#ifndef SYSFS_H
+#define SYSFS_H
+
+void sysfs_init(void);
+
+#endif /* SYSFS_H */